requirements.txt
Appium-Python-Client==0.45
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
eyes-common==4.1.24
eyes-core==4.1.24
eyes-selenium==4.1.24
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
more-itertools==8.2.0
packaging==20.3
Pillow==6.2.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
tinycss2==0.6.1
transitions==0.6.9
ua-parser==0.8.0
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.1.0
